sl_mpc_mate

Module bip32

Structs

XPubKey

Extended public key

Enums

BIP32Error

Errors while performing keygen

Prefix

BIP32 version bytes

Functions

derive_child_pubkey

Derive a child public key from a parent public key and a parent chain code

derive_xpub

Derive the extended public key for a given derivation path and prefix

generate_key_id

Generate a key ID from a root public key and root chain code

get_finger_print

Get the fingerprint of the root public key

Type Aliases

KeyFingerPrint

4-byte Key fingerprint