sl_mpc_mate

Module coord

Modules

stats

Structs

BufferedMsgRelay

Feed

Future for the feed method.

MaybeFeed

MessageSendError

Traits

MessageRelayService

Relay

Sink

A Sink is a value into which other values can be sent, asynchronously.

SinkExt

An extension trait for Sinks that provides a variety of convenient combinator functions.

Stream

A stream of values produced asynchronously.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.